backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic == 2.*
cloudinary==1.36.0
httpx
jwt
